Rename : 공통 컴포넌트 경로 components 하위로 일괄 이동
