+++ src/api/meal.js
@@ -4,3 +4,8 @@ import { getRequest } from './instance'
 export const getMealBest = () => {
   return getRequest(`/food/best`)
 }
+
+// 식단 상세 조회
+export const getMealDetail = (params) => {
+  return getRequest(`/food/board/${params.boardId}`)
+}
