+++ src/components/index.js
@@ -1,4 +1,5 @@
 import Button from './Button'
+import BoardList from './BoardList'
 import ContentCard from './ContentCard'
 import Footer from './Footer'
 import Layout from './Layout'
@@ -8,4 +9,4 @@ import QuadImages from './QuadImages'
 import SectionTitle from './SectionTitle'
 import Sidebar from './Sidebar'
 
-export { Button, ContentCard, Footer, Layout, Navbar, PageTitle, QuadImages, SectionTitle, Sidebar }
+export { Button, BoardList, ContentCard, Footer, Layout, Navbar, PageTitle, QuadImages, SectionTitle, Sidebar }
