--- src/components/list/index.js
@@ -1,3 +1,0 @@
-import BoardList from './BoardList'
-
-export { BoardList }
